refactor: house the standard flavor's module kit beside its assembly
buildStandardModules and standardFlavor are two halves of one concept —
the standard flavor's module graph and the assembly that lowers it into a
ShellConfig. They lived in sibling singleton directories, src/composition/
and src/flavors/. Move the kit beside standard.dart under flavors/ and drop
the composition/ directory; the two stay separate files because DI wiring
and shell assembly are distinct altitudes.

Record the collapse as a consequence in ADR-003 §6, leaving the §1.2 #6
diagnosis as the historical motivation.
