+++ lib/flavors.dart
@@ -1,3 +1,3 @@
 export 'src/flavors/standard.dart' show standard, standardFlavor;
-export 'src/composition/standard_modules.dart'
+export 'src/flavors/standard_modules.dart'
     show buildStandardModules, StandardModules;
