+++ lib/src/flavors/standard_modules.dart
